refactor(providers): split copilot model discovery
